[modify] 기존 워커 삭제
