+++ small_app/sumgo_worker.py
@@ -1,5 +1,6 @@
 import os
 import datetime
+from small_app.dataclass.user_data import FilterData
 import time
 
 from typing import List, Dict, Generator
@@ -11,48 +12,20 @@ from pathlib import Path
 import selenium
 from selenium.webdriver.support import expected_conditions as EC
 
+from base.base_worker import BaseWorker
 from custom_element import CustomElement
 
 
-class SumgoWorker:
-
-    WAITTING_SECONDS = 15
+class SumgoWorker(BaseWorker):
     SELECTED_COUNT, DELETED_COUNT = 0, 0
 
     def __init__(
-        self,
-        target_url,
-        need_words: List[str] = [""],
-        ben_words: List[str] = [""],
-        delete_mode: bool = False,
+        self, conf: str = "conf.ini", words: FilterData = None, is_delete: bool = True,
     ):
-        self.target_url = target_url
-        self.driver: selenium = "chrome"
-        self.user_info = Config(os.path.join(os.getcwd(), "small_app/") + "conf.ini")
-        self.need_words: List[str] = need_words
-        self.ben_words: List[str] = ben_words
-        self.delete_mode = delete_mode
-
-    @property
-    def driver(self) -> selenium:
-        return self.__driver
-
-    @driver.setter
-    def driver(self, value):
-        if value == "chrome":
-            self.__driver = webdriver.Chrome("/usr/local/bin/chromedriver")
-            self.__driver.implicitly_wait(self.WAITTING_SECONDS)
-        print(f"{value} driver is ready")
-
-    @property
-    def user_info(self):
-        return self.__user_info
-
-    @user_info.setter
-    def user_info(self, value):
-        user = value.get_map("user")
-        print(f"login_info : {user}")
-        self.__user_info = user
+        super().__init__(conf)
+        self.need_words: List[str] = words.need_words
+        self.ben_words: List[str] = words.ben_words
+        self.is_delete = is_delete
 
     def get_elem_by_xpath(self, xpath: str) -> selenium.webdriver.remote.webelement.WebElement:
         try:
@@ -61,11 +34,33 @@ class SumgoWorker:
             print(f"[Error] {e}")
             return None
 
-    def execute(self):
-        self.driver.get(self.target_url)
+    def click(self, xpath: str):
+        elem = self.get_elem_by_xpath(xpath)
+        elem.click()
 
-        self.click_login()
-        self.type_email_password()
+    def go_to_login_page(self):
+        login_xpath = '//*[@id="app-header"]/div[3]/ul/li[4]/a'
+        self.click(login_xpath)
+
+    def type_login_information(self):
+        email_xpath = '//*[@id="__BVID__231"]'
+        email_input = self.get_elem_by_xpath(email_xpath)
+        email_input.send_keys(self.user_data.email)
+
+        pwd_xpath = '//*[@id="__BVID__233"]'
+        pwd_input = self.get_elem_by_xpath(pwd_xpath)
+        pwd_input.send_keys(self.user_data.password)
+
+        login_submit_elem = self.get_elem_by_xpath(
+            '//*[@id="app-body"]/div/div/form/div/div[4]/button'
+        )
+        login_submit_elem.click()
+
+    def execute(self):
+        self.driver.get(self.user_data.target_url)
+
+        self.go_to_login_page()
+        self.type_login_information()
 
         close_xpath = '//*[@id="__BVID__265___BV_modal_body_"]/div[1]'
         self.close_pop_up(close_xpath)
@@ -131,7 +126,7 @@ class SumgoWorker:
                 self.check_words(message)
                 self.make_log(message, "selected")
             except ValueError as v:
-                if self.delete_mode:
+                if self.is_delete:
                     self.delete_request_item(elem)
                 self.make_log(message)
 
@@ -168,33 +163,11 @@ class SumgoWorker:
     def close_pop_up(self, xpath: str):
         self.click(xpath)
 
-    def click_login(self):
-        login_xpath = '//*[@id="app-header"]/div[3]/ul/li[4]/a'
-        self.click(login_xpath)
-
-    def type_email_password(self):
-        email_xpath = '//*[@id="__BVID__231"]'
-        email_input = self.get_elem_by_xpath(email_xpath)
-        email_input.send_keys(self.__user_info.get("email", "empty"))
-
-        pwd_xpath = '//*[@id="__BVID__233"]'
-        pwd_input = self.get_elem_by_xpath(pwd_xpath)
-        pwd_input.send_keys(self.__user_info.get("password", "pwd"))
-
-        login_submit_elem = self.get_elem_by_xpath(
-            '//*[@id="app-body"]/div/div/form/div/div[4]/button'
-        )
-        login_submit_elem.click()
-
-    def click(self, xpath: str):
-        elem = self.get_elem_by_xpath(xpath)
-        elem.click()
-
 
 if __name__ == "__main__":
-    url = "https://soomgo.com/"
     need_words = ["파이썬", "python"]
     ben_words = ["자바", "Java", "C언어", "c언어", "딥러닝", "머신러닝"]
-    obj = SumgoWorker(url, need_words, ben_words, delete_mode=True)
+    words = FilterData(need_words=need_words, ben_words=ben_words)
 
-    obj.execute()
+    sw = SumgoWorker("conf.ini", words)
+    sw.execute()
